--- a/l5kit/l5kit/geometry/transform.py
+++ b/l5kit/l5kit/geometry/transform.py
@@ -92,9 +92,13 @@ def transform_points(points: np.ndarray, transf_matrix: np.ndarray) -> np.ndarra
     ), f"transf_matrix ({transf_matrix.shape}) should be a square matrix."
 
     if points.shape[1] not in [2, 3]:
-        raise AssertionError("Points input should be (N, 2) or (N, 3) shape, received {}".format(points.shape))
+        raise AssertionError(
+            f"Points input should be (N, 2) or (N, 3) shape, received {points.shape}"
+        )
 
-    assert points.shape[1] == transf_matrix.shape[1] - 1, "points dim should be one less than matrix dim"
+    assert points.shape[1] == transf_matrix.shape[1] - 1, (
+        "points dim should be one less than matrix dim"
+    )
 
     num_dims = len(transf_matrix) - 1
     transf_matrix = transf_matrix.T