My tests suggest that the modifications give a slight speed increase when used for 2-20000 points. Speed increase is larger when the points array is larger. For tests I purposly I used a continuous and non-continuous arrays with very similar results.

N = 20000
old: 168 µs ± 5.91 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
new: 159 µs ± 4.89 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

N = 20
old: 9.5 µs ± 139 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
new: 9.31 µs ± 173 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

I also modified the test for the above function as I found ways to bypass the test with "easy" solutions.